refactor (FinalFrameResult) : step3 - 매직넘버 제거 리팩토링
