+++ src/test/java/bowling/domain/frame/FinalNormalFrameResultTest.java
