+++ src/test/java/bowling/domain/frame/NormalFrameResultTest.java
